backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
google-genai>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
